+++ src/connection/cursor.ts
@@ -1,4 +1,3 @@
-// import _debug from 'debug'; // it is vulnerable
 import DoublyLinked from 'doublylinked';
 import { TaskQueue } from 'power-tasks';
 import { FieldInfo } from '../interfaces/field-info.js';
@@ -9,8 +8,6 @@ import { convertRowToObject } from '../util/convert-row-to-object.js';
 import { parseRow } from '../util/parse-row.js';
 import { Portal } from './portal.js';
 import { PreparedStatement } from './prepared-statement.js';
-
-const debug = (() => void 0) as any;// _debug("pgc:cursor");
 
 export class Cursor extends SafeEventEmitter {
   private readonly _statement: PreparedStatement;
@@ -35,7 +32,6 @@ export class Cursor extends SafeEventEmitter {
     this._parsers = parsers;
     this._queryOptions = queryOptions;
     this.fields = fields;
-    debug("[%s] constructor", this._portal.name);
   }
 
   get rowType(): "array" | "object" {
@@ -67,7 +63,6 @@ export class Cursor extends SafeEventEmitter {
 
   async close(): Promise<void> {
     if (this._closed) return;
-    debug("[%s] close", this._portal.name);
     await this._portal.close();
     await this._statement.close();
     this.emit("close");
@@ -79,7 +74,6 @@ export class Cursor extends SafeEventEmitter {
     const portal = this._portal;
     await this._taskQueue
         .enqueue(async () => {
-          debug("[%s] fetching rows", this._portal.name);
           const queryOptions = this._queryOptions;
           const r = await portal.execute(queryOptions.fetchCount || 100);
           if (r && r.rows && r.rows.length) {
@@ -94,7 +88,6 @@ export class Cursor extends SafeEventEmitter {
               }
             }
             this._rows.push(...r.rows);
-            debug("[%s] %d rows fetched", this._portal.name, r.rows.length);
             this.emit("fetch", r.rows);
           } else {
             await this.close();