+++ src/interfaces/script-execute-options.ts
@@ -23,5 +23,5 @@ export interface ScriptExecuteOptions extends DataMappingOptions {
    * When off (the default), a statement in a transaction block that generates an error aborts the entire transaction
    * @default true
    */
-  onErrorRollback?: boolean;
+  rollbackOnError?: boolean;
 }